renaming a service
